+++ src/context.ts
@@ -1,9 +1,12 @@
+import { Logger } from 'pino';
 import { DMPHubAPI } from './datasources/dmphub-api';
 import { MysqlDataSource } from './datasources/mysqlDB';
 
-export type DataSourceContext = {
+export type MyContext = {
+  logger: Logger;
+
   dataSources: {
     dmphubAPIDataSource: DMPHubAPI;
-    sqlDataSource: MysqlDataSource
+    sqlDataSource: MysqlDataSource;
   };
 }