+++ src/resolvers/dmsp.ts
@@ -1,13 +1,26 @@
+import { formatLogMessage } from '../logger';
 import { Resolvers } from "../types";
+
+function prepareLogger(logger, session = null) {
+
+}
 
 export const resolvers: Resolvers = {
   Query: {
     // returns a DMSP that matches the specified DMP ID
-    dmspById: (_, { dmspId }, { dataSources }) => {
+    dmspById: (_, { dmspId }, { logger, dataSources }) => {
+      const logMessage = `Resolving query dmspById(dmspId: '${dmspId}')`;
+
       return new Promise((resolve, reject) => {
         dataSources.dmphubAPIDataSource.getDMSP(encodeURIComponent(dmspId))
-          .then(rows => resolve(rows))
-          .catch(error => reject(error));
+          .then(rows => {
+            formatLogMessage(logger).debug(logMessage);
+            resolve(rows)
+          })
+          .catch(err => {
+            formatLogMessage(logger, { err, dmspId }).error(`ERROR: ${logMessage} - ${err.message}`);
+            reject(err)
+          });
       });
     },
   },